Project 1 - Finished Parser Beta
Finished initial version of parser, needs to be debuged and refine. Its
output is not yet linked to interface, need three simple functions to do
so. Created a generic DynamicTreeNode class for CST (and maybe future
projects). Only did basic version of Node, not overall Tree class, and
needs to be polished. This adds a new .dll to project by the way.