Project 1 - Lexer and Parser
Completed and hopefully I tested enough. Builds CST and basic symbol
table. After you hit compile, be sure to check the Lexer, Parser, and
Symbol table tabs.  My DynamicTreeNode class(and dll) works, but I need
to polish it , and add features to make it a component for use in future
projects. I will do that as the project comes along.