Implemented Requested Changes
I changed ParseBlock to not allow the first block to not
have braces, stoped repairing early EOF, and caused digits or capital
characters in string literals to be errors(they used to be warnings and
I ignored them). Actual keywords and IDs can still be a mix of upper and
lower case, it does not matter once a CST and Symbol table is built, it
wont effect op code generation.